enhance the integration test
